networking overhaul
Overhauled to implement networking. I moved the main method to
MainClient and MainServer so that we have separate starting methods, for
now. The client has a single DataPortClient object that connects to
server. The server has a ServerConnectionTable that holds a
DataPortServer for each connection. The server also has it's own
DataPortClient running for the local client. Made a myriad of other
changes to help simplify/strengthen our code. If you have any questions
let me know!

TO DEMO THE CODE: run MainServer, then run MainClient three times. You
will see the server send a message to the 4 clients (including itself),
and then the clients' responses. This does not exit cleanly if you
terminate the server first, as I haven't gone about closing the sockets
yet.