another networking overhaul
First off, the server now listens for connections in a separate thread
(so it keeps running stuff while listening) and can handle any number of
connections, no longer just 4.

Also, there's no more crazy freakout when a process is terminated.

There's just one kind of DataPort now, used for every host-to-host
connection to send and receive Strings. ServerConnectionTable represents
a bunch of DataPort connections (all the server's client connections).
Each DataPort in ServerConnectionTable has a uniqueId (might correlate
to the playerId later on) that can be used to identify the client.

To send on a DataPort, you simply use the dataport.send(msg) method.

To get incoming messages, create a LinkedList<String> variable (it's a
queue) and set it to the return of the dataport's getIncMessageQueue()
method. This queue is populated by the dataport's thread that's
constantly listening for messages. To see if there's a new message, use
queue.peek(). To pop it off the queue, use queue.poll().

You can look at MainClient and MainServer to see how setup is done as
well as how messages are sent and received.

Last note: I'm not sure why but I had to copy the source files into a
new project to get them to run. Eclipsed refused to see that there was a
main function. Perhaps a bug somewhere in the Eclipse project files in
the last commit?