Began working on a simulation
Started implementing a basic network play simulation.
The idea would be that the client is created and it waits for the server
to become active. Once the server is created they message each other and
when the client receives a message from the server successfully, it will
create a player which will then be drawn on it's GUI.

The simulation is not yet ready.