Began the long process of GUI commenting
Many of the methods and a large portion of the code was and still is in
a developing state. Comments were not added since code was often
rearranged, renamed, removed, relocated, etc.

Comments however are necessary and thus time needs to be spent on making
the code readable from a higher level.